fix: change types positions to prevent loop of import
refs: SHELL-128
